Rename get & delete hosts containers to get & delete instances
